:tada: NEW ARCHITECTURE INTEGRATED - MORE FRIENDLY + :tada: the logic behind the multilingual feature have been upgraded for a better SEO, includint the dynamic change of the lang attribute in the html tag, the referencement of the the different langues options the <Head> tag, as the integration of a logic directly looking in the client browser for its own settings in order to optimze its first experience + :tada: upgrade the import font logic
