+++ components/Profil/Profil.tsx
@@ -1,5 +1,5 @@
-import { useLanguage } from "@/app/context/LanguageContext";
-import { TranslationsType } from "@/app/types/TranslationsType";
+import { useLanguage } from "@/context/LanguageContext";
+import { TranslationsType } from "@/types/TranslationsType";
 import classes from "./Profil.module.scss";
 import Image from "next/image";
 import profilePicture from "../../public/images/profil_picture.png";
