+++ next.config.js
@@ -4,4 +4,8 @@ module.exports = {
   sassOptions: {
     includePaths: [path.join(__dirname, "styles")],
   },
+  i18n: {
+    locales: ["en", "fr"],
+    defaultLocale: "en",
+  },
 };
