+++ styles/globals.css
@@ -1,3 +1,5 @@
+@import url("https://fonts.googleapis.com/css2?family=Raleway:wght@400;600&family=Inter:wght@400;700&display=swap");
+
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
@@ -6,6 +8,7 @@
   --foreground-rgb: 0, 0, 0;
   --background-start-rgb: 214, 219, 220;
   --background-end-rgb: 255, 255, 255;
+  --font-primery: "Raleway", "Inter", sans-serif;
 }
 
 @media (prefers-color-scheme: dark) {
@@ -18,7 +21,9 @@
 
 body {
   color: rgb(var(--foreground-rgb));
+  /* ! TO TAKE OUT AT THE END */
   background: yellow;
+  font-family: var(--font-primery);
 }
 
 @layer utilities {
